Introduce linux kernel list API in quick sort
In the previous implementation of quick sort and related functions, we
use a self-defined structure "node_t" and implement its related
functions to do the list manipulation.

Now we introduce the linux kernel linked list which is also called
"struct list_head" in the project and replace all the related functions
by linux kernel list api to enforce a stronger consistancy and more
elegance implementation.