Contruct binary tree basic setup and test code
Setup the basic implementation of constructing binary tree in the c file
of tree and the header file of tree.

Testing code is implemented in the c file of main, we build a binary
tree of height "treeHeight" randomly and find its preorder sequence and
inorder sequence, use the sequences to build a binary tree using
"buildTree" in the header file of tree.

In the last phase of testing we compare whether the binary tree pointed
by "clone_root" can generate the same preorder sequence and inorder
sequence like the "root" did, if both of the sequences are the same then
we can say these two binary tree is the same.