LRU basic set up and testing program
Build up basic implementation of LRU cache manipulation function and
implement a testing program to test the result of the LRU cache in
"main".

We apply random sequence of put and get operation and check the most
recently used element in LRU cache to determine whether the operation is
done correctly.