Correctly move the node in hlist to the front
In the previous version of cache put, cache get operations, we didn't
move the "hlist_node" which was updated the most recently to the front
of the corresponding hash list.

We correct the operation in this change by moving the updated
"hlist_node" to the front the its hash list.