Comparison between a naive find method and hlist
Considered the method of finding the node index in the inorder array.
Orginally we use hash list which is constructed by the linux-style
"struct hlist_node" to record every node's index in the inorder array so
we can search for it.

However in the lack of concern of the chosen for hash function, the use
of "struct hlist_node" might be in vain, so we try to use a naive method
to search for the index of node in the inorder array and compare the cpu
clock taken by this method and the orginal method which use "struct
hlist_node".