Encapsulate hash function
Hash function for the search of hash table was written raw in each
functions such as lRUCacheGet(), which leads to difficulties when trying
to maintain or alter the hash function.

We encapsulate the hash function by written it into a function called
lRU_hash_slot() which can return the pointer to struct hlist_head
directly.